requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==2.3.7
boto3==1.40.64
python-dotenv==1.0.1
Pillow==10.4.0
requests==2.32.3
gunicorn==21.2.0
PyJWT==2.8.0
